requirements.txt
Django==1.4.3
distribute==0.6.45
dj-database-url==0.2.1
django-gcm==0.9.3
django-selenium==0.9.6
django-tastypie==0.9.15
mimeparse==0.1.3
mock==1.0.1
psycopg2==2.5
python-dateutil==2.1
readline==6.2.4.1
selenium==2.33.0
six==1.3.0
wsgiref==0.1.2
